scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
firebase-admin==6.4.0
pydantic==2.6.1
python-slugify==8.0.4
